minor pruning
