+++ Procfile
@@ -1,1 +1,2 @@
+release: ./submoduleUpdate.sh
 web: node __sapper__/build
