+++ tests.json
@@ -1,1 +1,1 @@
-[{"date":"10/19/2020","time":"13:16:54.165","ip":"199.21.137.7","uploadSpeed":"53.14","downloadSpeed":"10.44","ping":"10.33","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC"},{"date":"10/19/2020","time":"13:21:31.689","ip":"199.21.137.7","uploadSpeed":"67.76","downloadSpeed":"72.77","ping":"6.76","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC"},{"date":"10/19/2020","time":"13:24:23.648","ip":"199.21.137.7","uploadSpeed":"67.26","downloadSpeed":"77.60","ping":"7.70","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/19/2020","time":"13:25:8.60","ip":"199.21.137.7","uploadSpeed":"98.08","downloadSpeed":"92.80","ping":"5.85","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200},{"date":"10/19/2020","time":"13:31:26.85","ip":"199.21.137.7","uploadSpeed":"91.51","downloadSpeed":"79.07","ping":"6.64","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200},{"date":"10/19/2020","time":"13:31:56.989","ip":"199.21.137.7","uploadSpeed":"83.17","downloadSpeed":"79.23","ping":"5.22","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/19/2020","time":"13:38:31.82","ip":"199.21.137.7","uploadSpeed":"103.83","downloadSpeed":"80.65","ping":"6.96","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/20/2020","time":"14:26:4.217","ip":"199.21.137.7","uploadSpeed":"313.31","downloadSpeed":"245.17","ping":"6.05","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/20/2020","time":"14:26:39.679","ip":"199.21.137.7","uploadSpeed":"267.85","downloadSpeed":"260.58","ping":"8.60","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200}]
+[{"date":"10/19/2020","time":"13:16:54.165","ip":"199.21.137.7","uploadSpeed":"53.14","downloadSpeed":"10.44","ping":"10.33","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC"},{"date":"10/19/2020","time":"13:21:31.689","ip":"199.21.137.7","uploadSpeed":"67.76","downloadSpeed":"72.77","ping":"6.76","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC"},{"date":"10/19/2020","time":"13:24:23.648","ip":"199.21.137.7","uploadSpeed":"67.26","downloadSpeed":"77.60","ping":"7.70","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/19/2020","time":"13:25:8.60","ip":"199.21.137.7","uploadSpeed":"98.08","downloadSpeed":"92.80","ping":"5.85","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200},{"date":"10/19/2020","time":"13:31:26.85","ip":"199.21.137.7","uploadSpeed":"91.51","downloadSpeed":"79.07","ping":"6.64","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200},{"date":"10/19/2020","time":"13:31:56.989","ip":"199.21.137.7","uploadSpeed":"83.17","downloadSpeed":"79.23","ping":"5.22","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/19/2020","time":"13:38:31.82","ip":"199.21.137.7","uploadSpeed":"103.83","downloadSpeed":"80.65","ping":"6.96","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/20/2020","time":"14:26:4.217","ip":"199.21.137.7","uploadSpeed":"313.31","downloadSpeed":"245.17","ping":"6.05","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/20/2020","time":"14:26:39.679","ip":"199.21.137.7","uploadSpeed":"267.85","downloadSpeed":"260.58","ping":"8.60","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":200},{"date":"10/23/2020","time":"13:55:46.945","ip":"199.21.137.7","uploadSpeed":"269.18","downloadSpeed":"191.20","ping":"8.26","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/23/2020","time":"13:59:8.659","ip":"199.21.137.7","uploadSpeed":"283.72","downloadSpeed":"170.00","ping":"6.91","city":null,"latitude":null,"longitude":null,"isp":"AS22121 Topsham Communications LLC","chunkSize":100},{"date":"10/23/2020","time":"14:7:45.385","ip":"199.21.137.7","uploadSpeed":"285.38","downloadSpeed":"196.49","ping":"7.02","city":null,"latitude":null,"longitude":null,"isp":"","chunkSize":100},{"date":"10/23/2020","time":"14:9:24.502","ip":"199.21.137.7","uploadSpeed":"313.46","downloadSpeed":"189.63","ping":"8.05","city":null,"latitude":null,"longitude":null,"isp":"Topsham Communications LLC","chunkSize":100}]
